Struct reactor::Reactor

pub struct Reactor<'a> {
    // some fields omitted
}

Methods

impl<'a> Reactor<'a>

fn new() -> Reactor<'a>

Construct a new Reactor with (hopefully) intelligent defaults

fn configured(cfg: ReactorConfig) -> Reactor<'a>

Construct a new engine with defaults specified by the user

fn connect<'b>(&mut self, hostname: &'b str, port: usize, handler: Box<ConnHandler<'a>>) -> Result<Token>

Attempt a connection to the remote host specified at the remote hostname or ip address and the port. This is a connection on a non-blocking socket, so the connect call will return immediately. It requires a handler, to which it will supply a ConnResult which will indicate success or failure. On success, it will supply a socket, a token, and a remote IP addr. It then expects an Option> so that it can manage its events

fn listen<A: ToSocketAddrs>(&mut self, addr: A, handler: Box<ConnHandler<'a>>) -> Result<Token>

Listen on the supplied IP address:port for incoming TCP connections. This returns immediately and expects a handler to which it will supply ConnResult and expect Option> as a result

fn channel(&self) -> Sender<TaggedBuf>

fetch the event_loop channel for notifying the event_loop of new outbound data

fn timeout(&mut self, duration: u64, handler: Box<TimeoutHandler<'a>>) -> TimerResult<(Timeout, Token)>

Set a timeout to be executed by the event loop after duration milliseconds The supplied handler, which is a FnMut will be invoked no sooner than the timeout

fn timeout_conn(&mut self, duration: u64, ctxtok: Token) -> TimerResult<(Timeout, Token)>

Set a timeout to be executed by the event loop after duration milliseconds ctxtok specifies a Context to which the timer callback will be directed through the usual event dispatch mechanism for Contexts This is useful for handling protocols which have a ping/pong style timeout

fn register<C>(&mut self, ctx: C) -> Result<Token> where C: Context + 'static

Trade in an existing context (connected to a resource) and get a Token The context will be registered for whichever events are specified in its own interest retrieved by get_interest()

fn deregister(&mut self, token: Token) -> Result<Box<Context>>

Trade in your token for a Context and deregister the Context's socket/evented from the event_loop

fn run(&mut self)

process all incoming and outgoing events in a loop

fn run_once(&mut self)

process all incoming and outgoing events in a loop

fn shutdown(&mut self)

calculates the 11th digit of pi